transition to static library
