+++ CMakeLists.txt
@@ -1,0 +1,11 @@
+cmake_minimum_required(VERSION 3.30)
+project(fenster VERSION 1.0.0)
+
+set(CMAKE_CXX_STANDARD 20)
+
+include_directories(include)
+
+add_library(fenster
+        STATIC
+        src/fenster.cpp
+)
